Migrate more tests from jpa/test to orm/test/jpa
