--- hibernate-core/src/test/java/org/hibernate/jpa/test/mapping/ColumnWithExplicitReferenceToPrimaryTableTest.java
@@ -1,59 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.jpa.test.mapping;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import org.hibernate.cfg.AvailableSettings;
-import org.hibernate.jpa.boot.spi.Bootstrap;
-import org.hibernate.jpa.test.PersistenceUnitDescriptorAdapter;
-
-import org.junit.Test;
-
-import org.hibernate.testing.TestForIssue;
-import org.hibernate.testing.junit4.BaseUnitTestCase;
-
-/**
- * @author Steve Ebersole
- */
-public class ColumnWithExplicitReferenceToPrimaryTableTest extends BaseUnitTestCase {
-	@Test
-	@TestForIssue( jiraKey = "HHH-8539" )
-	public void testColumnAnnotationWithExplicitReferenceToPrimaryTable() {
-		final PersistenceUnitDescriptorAdapter pu = new PersistenceUnitDescriptorAdapter() {
-			@Override
-			public List<String> getManagedClassNames() {
-				return Arrays.asList( AnEntity.class.getName() );
-			}
-		};
-
-
-		final Map settings = new HashMap();
-		settings.put( AvailableSettings.HBM2DDL_AUTO, "create-drop" );
-
-		EntityManagerFactory emf = Bootstrap.getEntityManagerFactoryBuilder( pu, settings ).build();
-		emf.close();
-	}
-
-	@Entity
-	@Table( name = "THE_TABLE" )
-	public static class AnEntity {
-		@Id
-		public Integer id;
-		@Column( name = "THE_COLUMN", table = "THE_TABLE" )
-		public String theValue;
-	}
-}
