--- hibernate-core/src/test/java/org/hibernate/jpa/test/metagen/mappedsuperclass/attribute/AbstractNameable.java
@@ -1,34 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.jpa.test.metagen.mappedsuperclass.attribute;
-
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
-
-/**
- * @author Steve Ebersole
- */
-@MappedSuperclass
-public abstract class AbstractNameable {
-	private String name;
-
-	protected AbstractNameable() {
-	}
-
-	protected AbstractNameable(String name) {
-		this.name = name;
-	}
-
-	@Column(name = "name")
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-}
