+++ hibernate-core/src/test/resources/org/hibernate/orm/test/jpa/mapping/orm.xml
